gateway/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Temporal
temporalio==1.7.0

# Utilities
python-dotenv==1.0.0
pydantic==2.8.0

# Testing
pytest==8.0.0
httpx==0.27.0
